order by primary key if no field specified
