+++ README.md
@@ -41,7 +41,6 @@ Website](https://sqliter.grantramsay.dev)
 - [Quick Start](#quick-start)
 - [Contributing](#contributing)
 - [License](#license)
-- [Acknowledgements](#acknowledgements)
 
 ## Features
 
@@ -167,10 +166,3 @@ DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 OR OTHER DEALINGS IN THE SOFTWARE.
 ```
-
-## Acknowledgements
-
-SQLiter was initially developed as an experiment to see how helpful ChatGPT and
-Claud AI can be to speed up the development process. The initial version of the
-code was generated by ChatGPT, with subsequent manual/AI refinements and
-improvements.
