+++ tests/test_query.py
@@ -416,169 +416,6 @@ class TestQuery:
         assert results[0].name == "Jane Doe"
         assert results[1].name == "Jim Doe"
 
-    def test_order_default(self, db_mock) -> None:
-        """Test that the order method works as expected."""
-
-        # Define a simple model for the test
-        class OrderTestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name: str = "order_test_table"
-
-        # Create the table and insert records
-        db_mock.create_table(OrderTestModel)
-        db_mock.insert(OrderTestModel(name="John Doe"))
-        db_mock.insert(OrderTestModel(name="Jane Doe"))
-        db_mock.insert(OrderTestModel(name="Jim Doe"))
-
-        # Perform a query with ordering by name DESC
-        results = db_mock.select(OrderTestModel).order("name").fetch_all()
-
-        # Assert that the ordering works in descending order
-        assert len(results) == 3
-        assert results[0].name == "Jane Doe"
-        assert results[1].name == "Jim Doe"
-        assert results[2].name == "John Doe"
-
-    def test_order_with_reverse_false(self, db_mock) -> None:
-        """Test the order method works with reverse=False (ascending order)."""
-
-        class TestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name = "test_table"
-
-        db_mock.create_table(TestModel)
-        db_mock.insert(TestModel(name="Charlie"))
-        db_mock.insert(TestModel(name="Alice"))
-        db_mock.insert(TestModel(name="Bob"))
-
-        # Ascending order
-        results = (
-            db_mock.select(TestModel).order("name", reverse=False).fetch_all()
-        )
-        assert len(results) == 3
-        assert results[0].name == "Alice"
-        assert results[1].name == "Bob"
-        assert results[2].name == "Charlie"
-
-    def test_order_with_reverse_true(self, db_mock) -> None:
-        """Test the order method works with reverse=True (descending order)."""
-
-        class TestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name = "test_table"
-
-        db_mock.create_table(TestModel)
-        db_mock.insert(TestModel(name="Charlie"))
-        db_mock.insert(TestModel(name="Alice"))
-        db_mock.insert(TestModel(name="Bob"))
-
-        # Descending order
-        results = (
-            db_mock.select(TestModel).order("name", reverse=True).fetch_all()
-        )
-        assert len(results) == 3
-        assert results[0].name == "Charlie"
-        assert results[1].name == "Bob"
-        assert results[2].name == "Alice"
-
-    def test_order_direction_ascending(self, db_mock) -> None:
-        """Test that the order method works as expected when ASC specified."""
-
-        # Define a simple model for the test
-        class OrderTestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name: str = "order_test_table"
-
-        # Create the table and insert records
-        db_mock.create_table(OrderTestModel)
-        db_mock.insert(OrderTestModel(name="John Doe"))
-        db_mock.insert(OrderTestModel(name="Jane Doe"))
-        db_mock.insert(OrderTestModel(name="Jim Doe"))
-
-        # Perform a query with ordering by name DESC
-        results = (
-            db_mock.select(OrderTestModel)
-            .order("name", direction="asc")
-            .fetch_all()
-        )
-
-        # Assert that the ordering works in descending order
-        assert len(results) == 3
-        assert results[0].name == "Jane Doe"
-        assert results[1].name == "Jim Doe"
-        assert results[2].name == "John Doe"
-
-    def test_order_direction_desc(self, db_mock) -> None:
-        """Test that the order method works as expected descending."""
-
-        # Define a simple model for the test
-        class OrderTestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name: str = "order_test_table"
-
-        # Create the table and insert records
-        db_mock.create_table(OrderTestModel)
-        db_mock.insert(OrderTestModel(name="John Doe"))
-        db_mock.insert(OrderTestModel(name="Jane Doe"))
-        db_mock.insert(OrderTestModel(name="Jim Doe"))
-
-        # Perform a query with ordering by name DESC
-        results = (
-            db_mock.select(OrderTestModel)
-            .order("name", direction="desc")
-            .fetch_all()
-        )
-
-        # Assert that the ordering works in descending order
-        assert len(results) == 3
-        assert results[0].name == "John Doe"
-        assert results[1].name == "Jim Doe"
-        assert results[2].name == "Jane Doe"
-
-    def test_order_with_both_reverse_and_direction(self, db_mock) -> None:
-        """Test both 'direction' and 'reverse' raises an InvalidOrderError."""
-
-        class TestModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name = "test_table"
-
-        db_mock.create_table(TestModel)
-        db_mock.insert(TestModel(name="Charlie"))
-        db_mock.insert(TestModel(name="Alice"))
-        db_mock.insert(TestModel(name="Bob"))
-
-        # Both 'reverse' and 'direction' specified (should raise error)
-        with pytest.raises(
-            InvalidOrderError,
-            match="Cannot specify both 'direction' and 'reverse'",
-        ):
-            db_mock.select(TestModel).order(
-                "name", direction="ASC", reverse=True
-            )
-
-    def test_order_deprecation_warning(self, db_mock) -> None:
-        """Test that using 'direction' raises a DeprecationWarning."""
-
-        class TestModel(BaseDBModel):
-            name: str
-
-        with pytest.warns(
-            DeprecationWarning, match="'direction' argument is deprecated"
-        ):
-            db_mock.select(TestModel).order("name", direction="ASC")
-
     def test_limit_offset_order_combined(self, db_mock) -> None:
         """Test that limit, offset, and order can work together."""
 
@@ -609,49 +446,6 @@ class TestQuery:
         assert len(results) == 2
         assert results[0].name == "Jane Doe"
         assert results[1].name == "Jim Doe"
-
-    def test_order_invalid_direction(self, db_mock) -> None:
-        """Test that an invalid order direction raises an exception."""
-
-        # Define a simple model for the test
-        class InvalidOrderModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name: str = "invalid_order_test_table"
-
-        # Create the table for the model
-        db_mock.create_table(InvalidOrderModel)
-
-        # Attempt to order by an invalid direction
-        with pytest.raises(InvalidOrderError) as exc:
-            db_mock.select(InvalidOrderModel).order("name", direction="invalid")
-
-        assert (
-            "Invalid order value - 'invalid' is not a valid sorting direction"
-            in str(exc.value)
-        )
-
-    def test_order_invalid_field(self, db_mock) -> None:
-        """Test that an invalid order field raises an exception."""
-
-        # Define a simple model for the test
-        class InvalidOrderModel(BaseDBModel):
-            name: str
-
-            class Meta:
-                table_name: str = "invalid_order_test_table"
-
-        # Create the table for the model
-        db_mock.create_table(InvalidOrderModel)
-
-        # Attempt to order by an invalid field
-        with pytest.raises(InvalidOrderError) as exc:
-            db_mock.select(InvalidOrderModel).order("invalid_field")
-
-        assert "Invalid order value - 'invalid_field' does not exist" in str(
-            exc.value
-        )
 
     def test_limit_edge_cases(self, db_mock) -> None:
         """Test limit with edge cases like zero and negative values."""
